backend/requirements.txt
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
sqlparse
python-dotenv
psycopg2-binary
pytz
python-dateutil
openai
gunicorn
whitenoise